修正int錯誤

+++ HealthMgr/BmiCalculator.cs
@@ -17,7 +17,7 @@
         {
             float result = 0;
             //將float改成int試試看
-            float height = (int)Height / 100;
+            float height = (float)Height / 100;
             result = Weight / (height * height);
 
             return result;
